Use GPP as preprocessor instead of C compiler
The GCC preprocessor (4.8+) adds some C-style comments (/* */) to
the generated html files. This commit defines GPP as the default
preprocessor in the Makefile. GPP ('a general-purpose preprocessor')
works as a drop-in replacement and is better suited for the task.

You still can use GCC (or another C compiler) with
make PP='gcc -E -P -C -nostdinc'

GPP project page and repository:
https://logological.org/gpp
https://github.com/logological/gpp